Add support for the Allwinner A20 Display-Engine
Add compatible string for Allwinner A20 devices.
Seems the Allwinner A20 is also capable of using the sun4i drm module.

Altered the device-tree for the sun7i-a20 device.
This allows the sun4i-drm driver to find the display-engine.
Removed the front-end 1, back-end 1, and tcon 1. As I personally don't
need them. I know there is code commited which alters all device trees
and it mainlined already. This code also reworks the sun4i-drm driver.
So the support for the removed nodes will be added back in at that point.

Signed-off-by: Thomas van Kleef Vitsch Electronics <linux-dev@vitsch.nl>