Use different clock driver on the Allwinner A20.
Use the different clockdriver for PLL6. This driver allows us to set a
required clock rate for the TCON CH0 which needs high clockrates for LVDS
panels connected to it.

Remove pll7x2 from tcon0_ch0 clock node in sun7i-a20.dtsi. This is not a valid
clock and needs to be replaced with pll6. This clock is multiplied by 2 for
this module.

PLL6 was used a clock source for the CPU, propebly to support frequency scaling.
Removed PLL6 as a clock source for the CPU, this makes the system use PLL1.
PLL1 is also called CORE clock which suggests it was intended for the CPU.

AHB bus also used the PLL6, this now uses the AXI clock again.

Signed-off-by: Thomas van Kleef <linux-dev@vitsch.nl>